Refactoring and use state in UrlWhiteListFormModal
* use state instead of props to get the urlwhitelist
* add UrlWhitelistInput and URLWhitelistFormModal to index.js in
component/common
* Fix linter
